Phase F-1 wave 4: extract Task mixins (§3.7 fallback split) + MRO guard (#114)
Extract every remaining task op off `_StoreBase`. The combined task
mixin measured 955 SLOC / MI 12.96 — over the 800 file gate — so the
plan §3.7 fallback fires and the task ops split into two mixins:

- `_ops/tasks_create.py` — `_TaskCreateOpsMixin` (create_task, the
  typed create_*_task builders, the `_insert_*` composite-commit
  bodies, the §6.4 at-most-one-live guards; 258 SLOC).
- `_ops/tasks_lifecycle.py` — `_TaskLifecycleOpsMixin` (task reads,
  claim/submit/accept/reject/reclaim/reassign, the per-role
  accept/reject helpers, the submit-gate + acceptance validators;
  715 SLOC).

The create/lifecycle seam is clean: create methods call only
`_StoreCore` primitives + the create-side live-task guards; lifecycle
never calls create-side helpers — no cross-mixin pyright coupling.

- `_base.py` is now a 235-SLOC core: constants, `_Tx`, `_StoreCore`,
  the mixin imports, and the composite `_StoreBase`.
- Composite is now 8 mixins + `_StoreCore`; a module-load-time MRO
  assertion (`__mro__[1:10]`, plan §6/§8.1) fails loud on any future
  bases reorder.
- L-O (`reassign_task`): extracted verbatim. It already measures 67
  LEN / CC 6 — inside the plan §D.6 target band (60-80) thanks to the
  pre-existing `_stage_reassign_reclaim` / `_reassign_event_payload`
  helpers. §D.6 explicitly warns against over-splitting an
  already-factored method for the LEN gate, so no cosmetic
  `_reassign_pending_or_claimed` helper was added; the deliverable's
  goal (under threshold) is already met. (Deviation surfaced for
  review.)
- `_validated_update` re-export: `_base` no longer uses the helper, so
  ruff prunes the wave-1 re-export. Rather than keep a dead
  `# noqa: F401` shim (BASE.md: no backwards-compat shims in
  greenfield), the 4 in-repo eden-git test import sites now point at
  the helper's canonical home `eden_storage._ops._helpers`. (Deviation
  from the plan's wave-1 re-export note; surfaced for review.)

Behavior-preserving: storage suite 373 passed / 185 skipped, eden-git
integrator tests 38 passed, full pyright clean, complexity gate green
(0 blocking; `_base.py`'s now-redundant annotation removed in wave 5).
